requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
bloxroute-cli==1.99.12
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.5
cytoolz==0.11.0
dataclasses==0.6
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.2
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
orjson==3.4.6
parsimonious==0.8.1
protobuf==3.18.0
pycryptodome==3.10.1
pyhumps==1.6.1
PyJWT==1.7.1
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rlp==2.0.1
six==1.16.0
toolz==0.11.1
twilio==6.63.2
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
web3==5.13.1
websockets==8.1
yarl==1.6.3
